+++ neural_lam/train_model.py
@@ -5,6 +5,7 @@ import time
 from argparse import ArgumentParser
 
 # Third-party
+# for logging the model:
 import pytorch_lightning as pl
 import torch
 from lightning_fabric.utilities import seed
@@ -248,10 +249,17 @@ def main(input_args=None):
 
     # Logger Settings
     parser.add_argument(
-        "--wandb_project",
+        "--logger",
+        type=str,
+        default="wandb",
+        choices=["wandb", "mlflow"],
+        help="Logger to use for training (wandb/mlflow) (default: wandb)",
+    )
+    parser.add_argument(
+        "--logger-project",
         type=str,
         default="neural_lam",
-        help="Wandb project name (default: neural_lam)",
+        help="Logger project name, for eg. Wandb (default: neural_lam)",
     )
     parser.add_argument(
         "--val_steps_to_log",
@@ -443,10 +451,18 @@ def main(input_args=None):
         )
     )
 
-    logger = pl.loggers.WandbLogger(
-        project=args.wandb_project,
-        name=run_name,
-        config=dict(training=vars(args), datastore=datastore._config),
+
+
+    training_logger = utils.setup_training_logger(
+        datastore=datastore, args=args, run_name=run_name
+    )
+
+    checkpoint_callback = pl.callbacks.ModelCheckpoint(
+        dirpath=f"saved_models/{run_name}",
+        filename="min_val_loss",
+        monitor="val_mean_loss",
+        mode="min",
+        save_last=True,
     )
     trainer = pl.Trainer(
         max_epochs=args.epochs,
@@ -455,7 +471,7 @@ def main(input_args=None):
         accelerator=device_name,
         num_nodes=args.num_nodes,
         devices=devices,
-        logger=logger,
+        logger=training_logger,
         log_every_n_steps=1,
         callbacks=callbacks,
         check_val_every_n_epoch=args.val_interval,
@@ -465,11 +481,15 @@ def main(input_args=None):
 
     # Only init once, on rank 0 only
     if trainer.global_rank == 0:
-        utils.init_wandb_metrics(
-            logger, val_steps=args.val_steps_to_log
-        )  # Do after wandb.init
+        utils.init_training_logger_metrics(
+            training_logger, val_steps=args.val_steps_to_log
+        )  # Do after initializing logger
     if args.eval:
-        trainer.test(model=model, datamodule=data_module, ckpt_path=args.load)
+        trainer.test(
+            model=model,
+            datamodule=data_module,
+            ckpt_path=args.load,
+        )
     else:
         # Only feed fit method with checkpoint path if restore_opt
         ckpt_for_fit = args.load if args.restore_opt else None